Moved ScapulothoracicJoint to SimbodyEngine with other joints
