+++ OpenSim/Simulation/SimbodyEngine/ScapulothoracicJoint.cpp
