+++ OpenSim/Simulation/SimbodyEngine/ScapulothoracicJoint.h
